% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmatrix_to_dagitty.R
\name{adjmatrix_to_dagitty}
\alias{adjmatrix_to_dagitty}
\title{Convert an adjacency matrix to a dagitty object}
\usage{
adjmatrix_to_dagitty(adjmatrix)
}
\arguments{
\item{adjmatrix}{A named adjacency matrix where a 1 in the \code{i,j}
position denotes an edge from node \code{i} to node \code{j}.
The matrix row and column names should denote the node names.}
}
\value{
A \code{dagitty} object with the same edge orientations.
}
\description{
\code{adjmatrix_to_dagitty} takes as
input an adjacency matrix and converts it to
a \code{dagitty} object.
}
\examples{
library(orientDAG)
library(dagitty)
library(bnlearn)
library(igraph)
g <- dagitty("dag {
  bn [pos=\"0,0\"]
  dagitty [pos=\"0.5,-1\"]
  adjmatrix [pos=\"1,0\"]
  adjmatrix -> dagitty
  dagitty -> adjmatrix
  bn -> adjmatrix
}")

# Below we can see the conversion directions implemented in this package:
plot(g)

# dagitty -> adjmatrix
adjmatrix <- dagitty_to_adjmatrix(g)
print(adjmatrix)

# adjmatrix -> dagitty
dagitty_obj <- adjmatrix_to_dagitty(adjmatrix)
plot(graphLayout(dagitty_obj))

# bn -> adjmatrix
bn_obj <- empty.graph(c("adjmatrix", "bn", "dagitty"))
amat(bn_obj) <- adjmatrix
plot(bn_obj)

adjmatrix <- bn_to_adjmatrix(bn_obj)
plot(graph_from_adjacency_matrix(adjmatrix))
}
\seealso{
\code{\link{dagitty_to_adjmatrix}} and \code{\link{bn_to_adjmatrix}}
}
