% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagitty_to_adjmatrix.R
\name{dagitty_to_adjmatrix}
\alias{dagitty_to_adjmatrix}
\title{Convert a dagitty object to an adjacency matrix}
\usage{
dagitty_to_adjmatrix(daggity_obj)
}
\arguments{
\item{daggity_obj}{An object of class \code{dagitty}.}
}
\value{
A named adjacency matrix where a 1 in the \code{i,j}
position denotes an arrow from node \code{i} to node \code{j}.
The matrix row and column names denote the node names.
}
\description{
\code{dagitty_to_adjmatrix} takes as
input a \code{dagitty} object and converts it to
an adjacency matrix.
}
\examples{
library(orientDAG)
library(dagitty)
library(bnlearn)
library(igraph)
g <- dagitty("dag {
  bn [pos=\\"0,0\\"]
  dagitty [pos=\\"0.5,-1\\"]
  adjmatrix [pos=\\"1,0\\"]
  adjmatrix -> dagitty
  dagitty -> adjmatrix
  bn -> adjmatrix
}")

# Below we can see the conversion directions implemented in this package:
plot(g)

# dagitty -> adjmatrix
adjmatrix <- dagitty_to_adjmatrix(g)
print(adjmatrix)

# adjmatrix -> dagitty
dagitty_obj <- adjmatrix_to_dagitty(adjmatrix)
plot(graphLayout(dagitty_obj))

# bn -> adjmatrix
bn_obj <- empty.graph(c("adjmatrix", "bn", "dagitty"))
amat(bn_obj) <- adjmatrix
plot(bn_obj)

adjmatrix <- bn_to_adjmatrix(bn_obj)
plot(graph_from_adjacency_matrix(adjmatrix))
}
\seealso{
\code{\link{adjmatrix_to_dagitty}} and \code{\link{bn_to_adjmatrix}}
}
