% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient_dag.R
\name{orient_dag}
\alias{orient_dag}
\title{Orient DAG edges}
\usage{
orient_dag(adjmatrix, x, max_continuous_pairs_sample = 5000)
}
\arguments{
\item{adjmatrix}{An adjacency matrix representing the input DAG. Edges can be both oriented
or bi-directional.}

\item{x}{A data.frame with columns corresponding to nodes in input DAG}

\item{max_continuous_pairs_sample}{Estimating the causal direction between two
numeric variables can be costly in time. For this reason one can cap the number
of measurements used for it using this argument. A value of NULL indicates not to re-orient
edges between 2 continuous variables.}

\item{orient_continuous}{}
}
\value{
A string denoting whether \code{vec_1} causes \code{vec_2} or vice versa
}
\description{
\code{orient_dag} learns from input data the correct input DAG
edge orientations.
}
\details{
Depending on the 2 variables encoding (each is either numeric or discrete)
a specific method is dispatched to determine the causal direction between them.
When the 2 variables are continuous, we can use several the general correlation measure
and related criteria by calling \code{\link[generalCorr]{some0pairs}}
(see also \href{https://cran.r-project.org/web/packages/generalCorr/vignettes/generalCorr-vignette.pdf}{Vinod 2017})

When the 2 variables are discrete, we can use the distance correlation measure by calling
\code{\link[energy]{dcor}} (see also {https://arxiv.org/pdf/1803.07712.pdf}{Liu and Chan 2016}).

When one of the variables is discrete, and the other is continuous we can discretisize the continuous
variable by calling \code{\link[infotheo]{discretize}} and use the method for two discrete variables.
}
